+++ fastapi/app/config.py
@@ -1,4 +1,5 @@
 from pydantic import BaseSettings
+from typing import Optional
 import logging
 from functools import lru_cache # read .env only once
 
@@ -22,8 +23,8 @@ class EnvSettings(BaseSettings):
   JWT_ALG: str = "HS256"
   JWT_SECRET: str = "".join(random.choice(string.ascii_letters) for i in range(32))
   REFRESH_TOKEN_SECRET: str = "".join(random.choice(string.ascii_letters) for i in range(64))
-  API_PORT=8000
-  SQLALCHEMY_DB_URL: str = "" # to be made a list
+  API_PORT: int = 8000
+  SQLALCHEMY_DB_URL: Optional[str] = "" # to be made a list
   CORS_ORIGINS: str = ""
   S3_ENDPOINT_URL: str = ""
   S3_ACCESS_ID: str = ""